+++ kraftwerk-api/src/main/java/fr/insee/kraftwerk/api/services/MainService.java
@@ -5,8 +5,11 @@ import fr.insee.kraftwerk.api.client.GenesisClient;
 import fr.insee.kraftwerk.api.configuration.ConfigProperties;
 import fr.insee.kraftwerk.api.configuration.MinioConfig;
 import fr.insee.kraftwerk.api.configuration.VaultConfig;
+import fr.insee.kraftwerk.api.dto.LastJsonExtractionDate;
 import fr.insee.kraftwerk.api.process.MainProcessing;
-import fr.insee.kraftwerk.api.process.MainProcessingGenesis;
+import fr.insee.kraftwerk.api.process.MainProcessingGenesisLegacy;
+import fr.insee.kraftwerk.api.process.MainProcessingGenesisNew;
+import fr.insee.kraftwerk.core.data.model.Mode;
 import fr.insee.kraftwerk.core.exceptions.KraftwerkException;
 import fr.insee.kraftwerk.core.utils.KraftwerkExecutionContext;
 import fr.insee.kraftwerk.core.utils.files.FileSystemImpl;
@@ -32,6 +35,7 @@ import org.springframework.web.bind.annotation.RestController;
 import java.io.IOException;
 import java.util.Map;
 
+import java.time.LocalDateTime;
 
 @RestController
 @Slf4j
@@ -98,10 +102,14 @@ public class MainService extends KraftwerkService {
 	}
 
 	/**
-	 * @author Adrien Marchal
+	 * @deprecated
+	 * We decided to switch from campaignId to questionnaireModelId to select data that are extracted.
+	 * One of the reasons is the possibility to face surveys that have multiple questionnaires for one campaign.
+	 * Since 3.4.1
 	 */
+	@Deprecated (since = "3.4.1", forRemoval = true)
 	@PutMapping(value = "/main/genesis")
-	@Operation(operationId = "mainGenesis", summary = "${summary.mainGenesis}", description = "${description.mainGenesis}")
+	@Operation(operationId = "mainGenesis", summary = "${summary.mainGenesis.deprecated}", description = "${description.mainGenesis.deprecated}")
 	public ResponseEntity<String> mainGenesis(
 			@Parameter(description = "${param.campaignId}", required = true, example = INDIRECTORY_EXAMPLE) @RequestBody String campaignId,
 			@Parameter(description = "${param.batchSize}") @RequestParam(value = "batchSize", defaultValue = "1000") int batchSize,
@@ -112,9 +120,26 @@ public class MainService extends KraftwerkService {
 		return runWithGenesis(campaignId, withDDI, withEncryption, batchSize, workersNumbers, workerId);
 	}
 
+	@PutMapping(value = "/main/genesis/by-questionnaire")
+	@Operation(operationId = "mainGenesisByQuestionnaireId", summary = "${summary.mainGenesis}", description = "${description.mainGenesis}")
+	public ResponseEntity<String> mainGenesisByQuestionnaireId(
+			@Parameter(description = "${param.questionnaireModelId}") @RequestParam(required = true) String questionnaireModelId,
+			@Parameter(description = "${param.dataMode}") @RequestParam(required = false) Mode dataMode,
+			@Parameter(description = "${param.batchSize}") @RequestParam(value = "batchSize", defaultValue = "1000") int batchSize,
+			@Parameter(description = "${param.withEncryption}") @RequestParam(value = "withEncryption", defaultValue = "false") boolean withEncryption) {
+		boolean withDDI = true;
+		return runWithGenesisByQuestionnaire(questionnaireModelId, withDDI, withEncryption, batchSize, dataMode);
+	}
 
+	/**
+	 * @deprecated
+	 * We decided to switch from campaignId to questionnaireModelId to select data that are extracted.
+	 * One of the reasons is the possibility to face surveys that have multiple questionnaires for one campaign.
+	 * Since 3.4.1
+	 */
+	@Deprecated (since = "3.4.1", forRemoval = true)
 	@PutMapping(value = "/main/genesis/lunatic-only")
-	@Operation(operationId = "mainGenesisLunaticOnly", summary = "${summary.mainGenesis}", description = "${description.mainGenesis}")
+	@Operation(operationId = "mainGenesisLunaticOnly", summary = "${summary.mainGenesis.lunatic.deprecated}", description = "${description.mainGenesis.lunatic.deprecated}")
 	public ResponseEntity<String> mainGenesisLunaticOnly(
 			@Parameter(description = "${param.campaignId}", required = true, example = INDIRECTORY_EXAMPLE) @RequestBody String campaignId,
 			@Parameter(description = "${param.batchSize}") @RequestParam(value = "batchSize", defaultValue = "1000") int batchSize,
@@ -126,30 +151,40 @@ public class MainService extends KraftwerkService {
 		return runWithGenesis(campaignId, withDDI, withEncryption, batchSize, workersNumbers, workerId);
 	}
 
+	@PutMapping(value = "/main/genesis/by-questionnaire/lunatic-only")
+	@Operation(operationId = "mainGenesisLunaticOnlyByQuestionnaire", summary = "${summary.mainGenesis}", description = "${description.mainGenesis}")
+	public ResponseEntity<String> mainGenesisLunaticOnlyByQuestionnaire(
+			@Parameter(description = "${param.questionnaireModelId}") @RequestParam String questionnaireModelId,
+			@Parameter(description = "${param.dataMode}") @RequestParam(required = false) Mode dataMode,
+			@Parameter(description = "${param.batchSize}") @RequestParam(value = "batchSize", defaultValue = "1000") int batchSize,
+			@Parameter(description = "${param.withEncryption}") @RequestParam(value = "withEncryption", defaultValue = "false") boolean withEncryption
+	) {
+		boolean withDDI = false;
+		return runWithGenesisByQuestionnaire(questionnaireModelId, withDDI, withEncryption, batchSize, dataMode);
+	}
 
-	//WIP : poc one interrogation
-	//Objective : give the json with all the data
 	@GetMapping(value ="/json")
 	@Operation(operationId = "jsonExtraction", summary = "", description ="")
 	public ResponseEntity<Object> jsonExtraction(
-			@Parameter(description = "${param.campaignId}", required = true, example = INDIRECTORY_EXAMPLE) @RequestParam String campaignId,
-			@Parameter(description = "${param.interrogationId}", required = true, example = INDIRECTORY_EXAMPLE) @RequestParam String interrogationId
-	){
+			@Parameter(description = "${param.questionnaireModelId}", required = true, example = INDIRECTORY_EXAMPLE) @RequestParam String questionnaireModelId,
+			@Parameter(description = "${param.dataMode}") @RequestParam(required = false) Mode dataMode,
+			@Parameter(description = "${param.batchSize}") @RequestParam(value = "batchSize", defaultValue = "1000") int batchSize,
+			@Parameter(description = "Extract since") @RequestParam(value = "sinceDate",required = false) LocalDateTime since
+			){
 		FileUtilsInterface fileUtilsInterface = getFileUtilsInterface();
 		boolean withDDI = true;
 		boolean withEncryption = false;
 
-		MainProcessingGenesis mpGenesis = getMainProcessingGenesis(withDDI, withEncryption, fileUtilsInterface);
-		Map<String,Object> results;
+		MainProcessingGenesisNew mpGenesis = getMainProcessingGenesisByQuestionnaire(withDDI, withEncryption, fileUtilsInterface);
 		try {
-			results = mpGenesis.runMainJson(campaignId, interrogationId);
+			mpGenesis.runMainJson(questionnaireModelId, batchSize, dataMode, since);
 			log.info("Data extracted");
 		} catch (KraftwerkException e) {
 			return ResponseEntity.status(e.getStatus()).body(e.getMessage());
 		} catch (IOException e) {
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
 		}
-		return ResponseEntity.ok(results);
+		return ResponseEntity.ok(String.format("Data extracted for questionnaireModelId %s",questionnaireModelId));
 	}
 
 	@NotNull
@@ -175,7 +210,7 @@ public class MainService extends KraftwerkService {
 	private ResponseEntity<String> runWithGenesis(String campaignId, boolean withDDI, boolean withEncryption, int batchSize, int workersNumbers, int workerId) {
 		FileUtilsInterface fileUtilsInterface = getFileUtilsInterface();
 
-		MainProcessingGenesis mpGenesis = getMainProcessingGenesis(withDDI, withEncryption, fileUtilsInterface);
+		MainProcessingGenesisLegacy mpGenesis = getMainProcessingGenesis(withDDI, withEncryption, fileUtilsInterface);
 
 		try {
 			mpGenesis.runMain(campaignId, batchSize, workersNumbers, workerId);
@@ -187,10 +222,27 @@ public class MainService extends KraftwerkService {
 		return ResponseEntity.ok(campaignId);
 	}
 
+	@NotNull
+	private ResponseEntity<String> runWithGenesisByQuestionnaire(String questionnaireModelId,  boolean withDDI, boolean withEncryption, int batchSize, Mode dataMode) {
+		FileUtilsInterface fileUtilsInterface = getFileUtilsInterface();
+
+		MainProcessingGenesisNew mpGenesis = getMainProcessingGenesisByQuestionnaire(withDDI, withEncryption, fileUtilsInterface);
+
+		try {
+			mpGenesis.runMain(questionnaireModelId, batchSize, dataMode);
+		} catch (KraftwerkException e) {
+			return ResponseEntity.status(e.getStatus()).body(e.getMessage());
+		} catch (IOException e) {
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
+		}
+		return ResponseEntity.ok(questionnaireModelId);
+	}
 
 
 
-	@NotNull MainProcessingGenesis getMainProcessingGenesis(boolean withDDI, boolean withEncryption, FileUtilsInterface fileUtilsInterface) {
+
+	@NotNull
+	MainProcessingGenesisLegacy getMainProcessingGenesis(boolean withDDI, boolean withEncryption, FileUtilsInterface fileUtilsInterface) {
 
 		KraftwerkExecutionContext kraftwerkExecutionContext = new KraftwerkExecutionContext(
 				null,
@@ -200,7 +252,26 @@ public class MainService extends KraftwerkService {
 				limitSize
 		);
 
-		return new MainProcessingGenesis(
+		return new MainProcessingGenesisLegacy(
+				configProperties,
+				new GenesisClient(new RestTemplateBuilder(), configProperties),
+				fileUtilsInterface,
+				kraftwerkExecutionContext
+		);
+	}
+
+	@NotNull
+	MainProcessingGenesisNew getMainProcessingGenesisByQuestionnaire(boolean withDDI, boolean withEncryption, FileUtilsInterface fileUtilsInterface) {
+
+		KraftwerkExecutionContext kraftwerkExecutionContext = new KraftwerkExecutionContext(
+				null,
+				false,
+				withDDI,
+				withEncryption,
+				limitSize
+		);
+
+		return new MainProcessingGenesisNew(
 				configProperties,
 				new GenesisClient(new RestTemplateBuilder(), configProperties),
 				fileUtilsInterface,