+++ kraftwerk-functional-tests/src/test/java/cucumber/functional_tests/GenesisDefinitions.java
@@ -6,7 +6,8 @@ import com.opencsv.CSVReader;
 import com.opencsv.CSVReaderBuilder;
 import com.opencsv.exceptions.CsvValidationException;
 import cucumber.TestConstants;
-import fr.insee.kraftwerk.api.process.MainProcessingGenesis;
+import fr.insee.kraftwerk.api.process.MainProcessingGenesisLegacy;
+import fr.insee.kraftwerk.api.process.MainProcessingGenesisNew;
 import fr.insee.kraftwerk.core.Constants;
 import fr.insee.kraftwerk.core.data.model.Mode;
 import fr.insee.kraftwerk.core.data.model.SurveyUnitUpdateLatest;
@@ -42,8 +43,8 @@ import static cucumber.functional_tests.MainDefinitions.outDirectory;
 
 public class GenesisDefinitions {
 
-    ConfigStub configStub = new ConfigStub();
-    GenesisClientStub genesisClientStub = new GenesisClientStub(configStub);
+    static ConfigStub configStub = new ConfigStub();
+    static GenesisClientStub genesisClientStub = new GenesisClientStub(configStub);
 
     private boolean isUsingEncryption;
     KraftwerkExecutionContext kraftwerkExecutionContext;
@@ -79,6 +80,22 @@ public class GenesisDefinitions {
         surveyUnitUpdateLatest.getCollectedVariables().add(variableModel);
     }
 
+    @Given("We have a collected variable {string} in a document with QuestionnaireModelId {string}, InterrogationId {string} " +
+            "with value {string}")
+    public void add_collected_variable_document_by_questionnaire(String variableName, String questionnaireModelId, String interrogationId,
+                                                String value) {
+
+        SurveyUnitUpdateLatest surveyUnitUpdateLatest = getSurveyUnitUpdateLatestByQuestionnaire(questionnaireModelId, interrogationId);
+
+        VariableModel variableModel = new VariableModel();
+        variableModel.setVarId(variableName);
+        variableModel.setIteration(1);
+        variableModel.setScope(Constants.ROOT_GROUP_NAME);
+        variableModel.setValue(value);
+
+        surveyUnitUpdateLatest.getCollectedVariables().add(variableModel);
+    }
+
     @Given("We have a collected variable {string} in a loop named {string} iteration {int} in a document with " +
             "CampaignId " +
             "{string}, " +
@@ -101,7 +118,47 @@ public class GenesisDefinitions {
         surveyUnitUpdateLatest.getCollectedVariables().add(variableModel);
     }
 
+    @Given("We have a collected variable {string} in a loop named {string} iteration {int} in a document with " +
+            "QuestionnaireModelId " +
+            "{string}, " +
+            "InterrogationId {string} with value {string}")
+    public void add_loop_collected_variable_document_by_questionnaire(  String variableName,
+                                                                        String loopName,
+                                                                        int iteration,
+                                                                        String questionnaireModelId,
+                                                                        String interrogationId,
+                                                                        String value) {
+
+        SurveyUnitUpdateLatest surveyUnitUpdateLatest = getSurveyUnitUpdateLatestByQuestionnaire(questionnaireModelId, interrogationId);
+
+        VariableModel variableModel = new VariableModel();
+        variableModel.setVarId(variableName);
+        variableModel.setScope(loopName);
+        variableModel.setIteration(iteration);
+        variableModel.setIdParent(Constants.ROOT_GROUP_NAME);
+        variableModel.setValue(value);
+
+        surveyUnitUpdateLatest.getCollectedVariables().add(variableModel);
+    }
+
     @Given("We have a external variable {string} in a document with CampaignId {string}, InterrogationId {string} " +
+            "with value {string}")
+    public void add_external_variable_document_by_questionnaire(String variableName,
+                                               String questionnaireModelId,
+                                               String interrogationId,
+                                               String value) {
+        SurveyUnitUpdateLatest surveyUnitUpdateLatest = getSurveyUnitUpdateLatestByQuestionnaire(questionnaireModelId, interrogationId);
+
+        VariableModel variableModel = new VariableModel();
+        variableModel.setVarId(variableName);
+        variableModel.setIteration(1);
+        variableModel.setScope(Constants.ROOT_GROUP_NAME);
+        variableModel.setValue(value);
+
+        surveyUnitUpdateLatest.getExternalVariables().add(variableModel);
+    }
+
+    @Given("We have a external variable {string} in a document with QuestionnaireModelId {string}, InterrogationId {string} " +
             "with value {string}")
     public void add_external_variable_document(String variableName,
                                                String campaignId,
@@ -142,6 +199,31 @@ public class GenesisDefinitions {
         surveyUnitUpdateLatest.getExternalVariables().add(variableModel);
     }
 
+    @Given("We have a external variable {string} in a loop named {string} iteration {int} in a document with " +
+            "QuestionnaireModelId " +
+            "{string}, " +
+            "InterrogationId {string} with value {string}")
+    public void add_loop_external_variable_document_by_questionnaire(   String variableName,
+                                                                        String loopName,
+                                                                        int iteration,
+                                                                        String questionnaireModelId,
+                                                                        String interrogationId,
+                                                                        String value
+    ) {
+
+        SurveyUnitUpdateLatest surveyUnitUpdateLatest = getSurveyUnitUpdateLatestByQuestionnaire(questionnaireModelId, interrogationId);
+
+        VariableModel variableModel = new VariableModel();
+        variableModel.setVarId(variableName);
+        variableModel.setScope(loopName);
+        variableModel.setIteration(iteration);
+        variableModel.setIdParent(Constants.ROOT_GROUP_NAME);
+        variableModel.setValue(value);
+
+        surveyUnitUpdateLatest.getExternalVariables().add(variableModel);
+    }
+
+
     private SurveyUnitUpdateLatest getSurveyUnitUpdateLatest(String campaignId, String interrogationId) {
         List<SurveyUnitUpdateLatest> mongoFiltered = genesisClientStub.getMongoStub().stream().filter(
                 surveyUnitUpdateLatest -> surveyUnitUpdateLatest.getCampaignId().equals(campaignId)
@@ -166,11 +248,36 @@ public class GenesisDefinitions {
         return surveyUnitUpdateLatest;
     }
 
+    private SurveyUnitUpdateLatest getSurveyUnitUpdateLatestByQuestionnaire(String questionnaireModelId, String interrogationId) {
+        List<SurveyUnitUpdateLatest> mongoFiltered = genesisClientStub.getMongoStub().stream().filter(
+                surveyUnitUpdateLatest -> surveyUnitUpdateLatest.getQuestionnaireId().equals(questionnaireModelId)
+                        && surveyUnitUpdateLatest.getInterrogationId().equals(interrogationId)
+        ).toList();
+
+        SurveyUnitUpdateLatest surveyUnitUpdateLatest;
+        if (mongoFiltered.isEmpty()) {
+            surveyUnitUpdateLatest = new SurveyUnitUpdateLatest();
+
+            surveyUnitUpdateLatest.setCampaignId(questionnaireModelId);
+            surveyUnitUpdateLatest.setQuestionnaireId(questionnaireModelId);
+            surveyUnitUpdateLatest.setInterrogationId(interrogationId);
+            surveyUnitUpdateLatest.setMode(Mode.WEB);
+
+            surveyUnitUpdateLatest.setCollectedVariables(new ArrayList<>());
+            surveyUnitUpdateLatest.setExternalVariables(new ArrayList<>());
+        } else {
+            surveyUnitUpdateLatest = mongoFiltered.getFirst();
+        }
+        genesisClientStub.getMongoStub().add(surveyUnitUpdateLatest);
+        return surveyUnitUpdateLatest;
+    }
+
     @Given("We want to encrypt output data at the end of genesis process")
     public void activateEncryption(){
         this.isUsingEncryption = true;
     }
 
+    // To delete when we remove historical genesis endpoints
     @When("We use the Genesis service with campaignId {string}")
     public void launch_genesis(String campaignId) throws IOException, KraftwerkException {
         configStub.setDefaultDirectory(TestConstants.FUNCTIONAL_TESTS_DIRECTORY);
@@ -179,13 +286,31 @@ public class GenesisDefinitions {
                 TestConstants.getKraftwerkExecutionContext(null, isUsingEncryption);
 
 
-        MainProcessingGenesis mainProcessingGenesis = new MainProcessingGenesis(
+        MainProcessingGenesisLegacy mainProcessingGenesisLegacy = new MainProcessingGenesisLegacy(
                 configStub,
                 genesisClientStub,
                 new FileSystemImpl(configStub.getDefaultDirectory()),
                 kraftwerkExecutionContext
         );
-        mainProcessingGenesis.runMain(campaignId,1000, 1, 1);
+        mainProcessingGenesisLegacy.runMain(campaignId,1000);
+        System.out.println();
+    }
+
+    @When("We use the Genesis service with questionnaireModelId {string}")
+    public void launch_genesis_by_questionnaire(String questionnaireModelId) throws IOException, KraftwerkException {
+        configStub.setDefaultDirectory(TestConstants.FUNCTIONAL_TESTS_DIRECTORY);
+
+        kraftwerkExecutionContext =
+                TestConstants.getKraftwerkExecutionContext(null, isUsingEncryption);
+
+
+        MainProcessingGenesisNew mainProcessingGenesisNew = new MainProcessingGenesisNew(
+                configStub,
+                genesisClientStub,
+                new FileSystemImpl(configStub.getDefaultDirectory()),
+                kraftwerkExecutionContext
+        );
+        mainProcessingGenesisNew.runMain(questionnaireModelId,1000,null);
         System.out.println();
     }